UPSTREAM: media: videobuf2-v4l2: refuse qbuf if queue uses requests or vv.
Check if the vb2 queue uses requests, and if so refuse to
add buffers that are not part of a request. Also check for
the reverse: a vb2 queue did not use requests, and an attempt
was made to queue a buffer to a request.

We might relax this in the future, but for now just return
-EPERM in that case.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 9ac8090fda77f072815c209c80fb230e89cda18c)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: Ic87133d96608c2ec76e88ef42deae3d2db678ec1