UPSTREAM: media: vb2: add init_buffer buffer op
We need to initialize the request_fd field in struct vb2_v4l2_buffer
to -1 instead of the default of 0. So we need to add a new op that
is called when struct vb2_v4l2_buffer is allocated.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 8e013700bc12806d80f31ebe360916987f0e03df)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: Id401a5746c23b5d33cb5b9f6e7653f46f16d0499