UPSTREAM: media: v4l2-ctrls: improve media_request_(un)lock_for_update
The request reference count was decreased again once a reference to the
request object was taken. Postpone this until we finished using the object.

In theory I think it is possible that the request_fd can be closed by
the application from another thread. In that case when request_put is
called the whole request would be freed.

It's highly unlikely, but let's just be safe and fix this potential
race condition.

Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>
Acked-by: Sakari Ailus <sakari.ailus@linux.intel.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit ffda0b4c2481b2a0faf634abccb1da0d0ba4eddb)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I076f2c1e0e8572b3ac3144c874ff85c64a15ee24