UPSTREAM: ASoC: dapm: Fix a couple uninitialized ret variables
Smatch complains that these variables could be uninitialized.  The first
one in snd_soc_dai_link_event() is probably a false positive, because
probably we know the lists are not empty.  I would normally ignore the
warning, but GCC complains here as well so I just silenced the warning.
The "ret" in snd_soc_dapm_new_dai() does need to be initialized or it
leads to a bogus dereference in the caller.

Fixes: 3bbf5d34fd4a ("ASoC: dapm: Move error handling to snd_soc_dapm_new_control_unlocked")
Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 2e558a8127de7b2ed3302f9adcf332ba3feeadb2)
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>

Change-Id: I4f75a9700acfa38376c469d469e7404e5d88fba5