YUKAWA: HACK: of: Changes to allow us to re-enable the overlay_mgr
The of core has been reworked and overlays are only supported
from detached flattened device trees.

Since the out of tree overlay_mgr driver uses embedded fragments
in the base dtb, it breaks with the new of core changes.

Thus this patch exports some internal functions for use with
the overlay_mgr and also removes some checks that will catch
as the overlay_mgr breaks some expectations.

Long term we need to move away from overlay_mgr (or work
upstream  with the of developers to add support - though they
have not been particularly keen to the general concept).

Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: Ifff0441dc1c63a110e62bf54b0f0fa671bce173f