FROMGIT: clk: meson: mpll: add init callback and regs
Until now (gx and axg), the mpll setting on boot (whatever the
bootloader) was good enough to generate a clean fractional division.

It is not the case on the g12a. While moving away from the vendor u-boot,
it was noticed the fractional part of the divider was no longer applied.
Like on the pll, some magic settings need to applied on the mpll
register.

This change adds the ability to do that on the mpll driver.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
(cherry picked from commit 19855c8276fec011afd8617a634eb62a29e6b871
 https://github.com/BayLibre/clk-meson v5.3/drivers)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I176281597be904d025cf08c9af3f2ec1717e95c8