FROMGIT: clk: meson: eeclk: add init regs
Like the PLL and MPLL, the controller may require some magic setting to
be applied on startup.

This is needed when the initial setting is not applied by the boot ROM.
The controller need to do it when the setting applies to several clock,
like all the MPLLs in the case of g12a.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
(cherry picked from commit 19a18d42bf557b8420a55e0fce7be5aec9f8ef8c
 https://github.com/BayLibre/clk-meson v5.3/drivers)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: Icf6fd295fcd16c0248e92a2d83786ecba315bbf7