FROMGIT: ASoC: meson: axg-card: set link name based on link node name
So far the link names of the axg sound card was derived from the cpu name
of the link. Since the dai link must be unique, it works as long as a
device does not provide more than one cpu dai. However, the 'tohdmitx'
does provide 2 dais used as cpu on codec-to-codec links

Instead of cpu name, use the node name of the dai link. DT already enforce
the uniqueness of this name

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Tested-by: Neil Armstrong <narmstrong@baylibre.com>
Tested-by: Kevin Hilman <khilman@baylibre.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 1b74211011eb064914b8155a77a8aaae61cd27eb
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-5.3)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: I69954b19910bbb8fd8c257eb8d0a3afeee54833a