UPSTREAM: media: v4l2-dev/ioctl: require non-zero device_caps, verify sane querycap results
Now that all V4L2 drivers set device_caps in struct video_device, we can
add a check for this to ensure all future drivers fill this in.

Also verify that when the querycap ioctl is called the driver didn't
mess with the device_caps value and that capabilities is a superset
of device_caps.

Signed-off-by: Hans Verkuil <hverkuil-cisco@xs4all.nl>
Reviewed-by: Sakari Ailus <sakari.ailus@linux.intel.com>
[hverkuil-cisco@xs4all.nl: fix too-long line]
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>
(cherry picked from commit 3c1350501c21db8e3b1a38d9e97db29694305c3b)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>

Change-Id: Id94ee2b6be57bd4be914d22cccb2fd0d820fefc2