BACKPORT: ASoC: codec2codec: deal with params when necessary
When there is an event on codec to codec dai_link, we only need to deal
with params if the event is SND_SOC_DAPM_PRE_PMU, when .hw_params() is
called. For the other events, it is useless.

Also, dealing with the codec to codec params just before calling
.hw_params() callbacks give change to either party on the link to alter
params content in .startup(), which might be useful in some cases

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>
Link: https://lore.kernel.org/r/20190725165949.29699-4-jbrunet@baylibre.com
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 3dcfb397dad2ad55bf50de3c5d5a57090d35a18a
 git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git for-5.4)

Change-Id: Icbbfedc01332dd9d79f6a60b498d171cb6c34844