YUKAWA: ASoC: meson: axg-card: setup platform node
Set the platform node to be the same as the cpu node. This prevent
ASoC from inserting the dummy platform codec which would override the
the stream parameters set by the axg-fifo drivers.

commit 1d7689892878 ("ASoC: soc-core: allow no Platform on dai_link")
is the best (and right) solution but it depends and important rework
which would actually make us pull all ASoC changes since.

This patch implement what was considered to be solution before that.
Please drop it when commit
1d7689892878 ("ASoC: soc-core: allow no Platform on dai_link") lands in
your tree.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>

Change-Id: I3071f4d2ced21c2247c333f6638759c5eed23815