FROMLIST: media: v4l2-mem2mem: mark DONE any OUTPUT queued buffer after CMD_STOP
If a decoder needs a minimal buffer count to be queued on it's CAPTURE
queue, if a CMD_STOP is sent after a STREAMON but before all the required
buffers are queued, it should comply to the drain sequence and mark the
last queued buffer with V4L2_BUF_FLAG_LAST and mark it done to be dequeued.

This introduces a v4l2-mem2mem ioctl decoder command to track the command
sent to the decoder, and should be called by the affected drivers.

Suggested-by: Hans Verkuil <hverkuil@xs4all.nl>
Suggested-by: Maxime Jourdan <mjourdan@baylibre.com>
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
(cherry picked from commit 2ba737e63eea9f137c90a017ddb05a0ac2f1cb99)
Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Link: https://patchwork.linuxtv.org/patch/60385/
Change-Id: Ieb5d55fe3627e930cd2efbd72878673b338559dd